chore: split shared.js into multiple constants files
