+++ app/renderer/actions/Inspector.js
@@ -3,7 +3,7 @@ import {v4 as UUID} from 'uuid';
 
 import i18n from '../../configs/i18next.config.renderer';
 import {SAVED_FRAMEWORK, SET_SAVED_GESTURES, getSetting, setSetting} from '../../shared/settings';
-import {APP_MODE, NATIVE_APP} from '../components/Inspector/shared';
+import {APP_MODE, NATIVE_APP} from '../constants/COMMON';
 import AppiumClient from '../lib/appium-client';
 import frameworks from '../lib/client-frameworks';
 import {getOptimalXPath, getSuggestedLocators} from '../utils/locator-generation';
