+++ app/renderer/components/Inspector/HighlighterRects.js
@@ -3,7 +3,8 @@ import React from 'react';
 import HighlighterCentroid from './HighlighterCentroid';
 import HighlighterRectForBounds from './HighlighterRectForBounds';
 import HighlighterRectForElem from './HighlighterRectForElem';
-import {RENDER_CENTROID_AS, parseCoordinates} from './shared';
+import {RENDER_CENTROID_AS} from '../../constants/SCREENSHOT';
+import {parseCoordinates} from '../../utils/other';
 
 const {CENTROID, OVERLAP, EXPAND} = RENDER_CENTROID_AS;
