+++ app/renderer/components/Inspector/Screenshot.js
@@ -3,13 +3,9 @@ import React, {useRef, useState} from 'react';
 
 import HighlighterRects from './HighlighterRects';
 import styles from './Inspector.css';
-import {
-  DEFAULT_SWIPE,
-  DEFAULT_TAP,
-  INTERACTION_MODE,
-  POINTER_TYPES,
-  SCREENSHOT_INTERACTION_MODE,
-} from './shared';
+import {INTERACTION_MODE} from '../../constants/COMMON';
+import {POINTER_TYPES} from '../../constants/GESTURES';
+import {DEFAULT_SWIPE, DEFAULT_TAP, SCREENSHOT_INTERACTION_MODE} from '../../constants/SCREENSHOT';
 
 const {POINTER_UP, POINTER_DOWN, PAUSE, POINTER_MOVE} = POINTER_TYPES;
 const {TAP, SELECT, SWIPE, TAP_SWIPE} = SCREENSHOT_INTERACTION_MODE;
