+++ app/renderer/constants/COMMON.js
@@ -1,0 +1,27 @@
+export const APP_MODE = {
+  NATIVE: 'native',
+  WEB_HYBRID: 'web_hybrid',
+};
+
+export const NATIVE_APP = 'NATIVE_APP';
+
+export const DRIVERS = {
+  UIAUTOMATOR2: 'uiautomator2',
+  ESPRESSO: 'espresso',
+  XCUITEST: 'xcuitest',
+  FLUTTER: 'flutter',
+  MAC2: 'mac2',
+  WINDOWS: 'windows',
+  CHROMIUM: 'chromium',
+  SAFARI: 'safari',
+  GECKO: 'gecko',
+};
+
+// Selected tab in the Session Inspector
+export const INTERACTION_MODE = {
+  SOURCE: 'source',
+  COMMANDS: 'commands',
+  GESTURES: 'gestures',
+  RECORDER: 'recorder',
+  SESSION_INFO: 'sessionInfo',
+};
