+++ app/renderer/constants/GESTURES.js
@@ -1,0 +1,6 @@
+export const POINTER_TYPES = {
+  POINTER_UP: 'pointerUp',
+  POINTER_DOWN: 'pointerDown',
+  PAUSE: 'pause',
+  POINTER_MOVE: 'pointerMove',
+};
